requirements.txt
pystray>=0.19.5
Pillow>=11.0
